+++ playwright/masonry/utils/getGridItems.mjs
@@ -1,0 +1,7 @@
+// @flow strict
+import selectors from './selectors.mjs';
+
+// $FlowExpectedError[unclear-type] flow-typed def for playwright is…lacking
+export default function getGridItems(page /*: Object */) /*: Promise<any> */ {
+  return page.locator(selectors.gridItem).all();
+}